# Fixed type-instability in Deque iterator