As reported in at least issue https://github.com/JuliaCollections/DataStructures.jl/issues/619, iterating over `Deque` was type unstable. This also therefore extends to `Queue` and `Stack` which are based on `Deque`. Running your benchmark (and a new `sum` benchmark) gives the following 75x speed-up of iteration.
Before:
```
traverse 10^7 integers:
    Vector:   elapsed =   0.0000s
    Deque:    elapsed =   0.3877s
sum 10^7 integers:
    Vector:   elapsed =   0.0044s
    Deque:    elapsed =   0.5728s
```
After:
```
traverse 10^7 integers:
    Vector:   elapsed =   0.0000s
    Deque:    elapsed =   0.0051s
sum 10^7 integers:
    Vector:   elapsed =   0.0036s
    Deque:    elapsed =   0.0085s
```